requirements.txt
Panda
Numpy
seaborn 
matplotlib
scikit-learn
dill
xgboost
catboost
#-e .